requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
load-dotenv==0.1.0
magic-filter==1.0.12
multidict==6.0.5
outcome==1.3.0.post0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PySocks==1.7.1
python-dotenv==1.0.1
selenium==4.20.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wsproto==1.2.0
yarl==1.9.4